Refactor ChatPeek to be in Chat
